requirements.txt
asgiref==3.5.2
attrs==21.4.0
Django==4.0.4
djangorestframework==3.13.1
drf-spectacular==0.22.1
inflection==0.5.1
jsonschema==4.5.1
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
simplejson==3.17.6
sqlparse==0.4.2
uritemplate==4.1.1
